Update to README.md
Adds example of basic implementation of use_secure_password and .authenticate methods. It seems like this implementation is becoming more and more common in the Rails community. To the best of my research, it looked like those methods have been available since Rails 4.0. If this is incorrect, please let me know, and I'll happily correct the reference so people don't try an incorrect usage.