BCrypt::Password comparison edge-case / bug
While working on the Sorcery rework, I found an edge case/bug where taking a
valid BCrypt::Password instance and trying to compare it using `==` against an
identical string causes it to return false.

This can be solved by checking if the secret being compared by `==` is a valid
BCrypt hash, and comparing it directly in those situations.

This solution will need to be vetted that it doesn't cause any vulnerabilities.